+++ src/index.ts
@@ -11,8 +11,10 @@ export {
 export type { EditTreeColumnsType } from './EditFormTreeTable/types';
 
 // 导出：基于dnd-kit的拖拽排序表格组件
-// export { default as DragSortSingleTable } from './DragSortSingleTable';
-// export type { DndKitSortTableProps } from './DragSortSingleTable/types';
+export { default as DragSortSingleTable } from './DragSortSingleTable';
+
+// 导出：基于dnd-kit的拖拽排序表格组件
+export { default as DragSortTreeTable } from './DragSortTreeTable';
 
 // 导出：ErrorBoundary组件
 export { default as ErrorBoundary, useErrorBoundary } from './ErrorBoundary';